채널 엔티티 성공

+++ src/main/java/hot/member/domain/ChCategory.java
@@ -23,7 +23,7 @@ public class ChCategory {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "ch_cate_no")
-	private Long chCateNo;
+	private Integer chCateNo;
 	
 	@Column(name = "ch_cate_name")
 	private String chCateName;
